Denis Volocaru - Added all of the entities from the database as objects.
Updated "PerformanceRepository" - to retrieve all performances from the database.

Updated "TicketSalesPage" - Created a seating plan for the Main Hall , with selectable seats. Once a seat is selected it is marked as green, and after another click it is marked as red (booked). Unselected seats are grey, while accesible seats are purple. (I still need to do database implementation,  in order to link these color codes according to the database, for example mark the seat as booked not only on the screen, but also on the database, and also retreive any booked seats from the db, and make sure to display them as red on the seating layout.